fix(dawn): make selected radio button visible
Due to a bug IDEA uses the checkbox background for the radio button but
there is no way to set the "inner dot" icon color which makes it very
hard for the light dawn theme to maintain proper visibility.
This seems like a good tradeoff to make it usable at least.